+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Models\RoleModel;
 use App\Models\UserModel;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use RealRashid\SweetAlert\Facades\Alert;
 use Yajra\DataTables\Facades\DataTables;
 
 class UserController extends Controller
@@ -92,8 +93,10 @@ class UserController extends Controller
         // Buat user baru
         UserModel::create($validatedData);
 
+        Alert::toast('Data pengguna berhasil ditambah', 'success');
+
         // Redirect ke halaman kelola pengguna
-        return redirect()->route('admin.kelolaPengguna.index')->with('success', 'Data berhasil ditambahkan!');
+        return redirect()->route('admin.kelolaPengguna.index');
     }
 
     public function show($id)
@@ -181,15 +184,25 @@ class UserController extends Controller
                 'foto' => $updateFoto['foto']
             ]);
         }
-        return redirect()->route('admin.kelolaPengguna.index')->with('success', 'Data Berhasil Diubah!');
+        Alert::toast('Data pengguna berhasil diubah', 'success');
+        return redirect()->route('admin.kelolaPengguna.index');
     }
 
     public function destroy($id) {
-        $user = UserModel::find($id);
-        if ($user->foto) {
-            Storage::disk('public')->delete($user->foto);
+        $check = UserModel::find($id);
+        if (!$check) {
+            Alert::toast('Data pengguna tidak ditemukan', 'error');
+            return redirect('/kelolaPengguna');
         }
-        UserModel::destroy($id);
-        return redirect()->route('admin.kelolaPengguna.index')->with('success', 'Data Berhasil Dihapus!');
+        try{
+            $kelolaPengguna = UserModel::find($id);
+            Storage::disk('public')->delete($kelolaPengguna->foto);
+            UserModel::destroy($id);
+            Alert::toast('Data pengguna berhasil dihapus', 'success');
+            return redirect('/kelolaPengguna');
+        }catch(\Illuminate\Database\QueryException $e){
+            Alert::toast('Data pengguna gagal dihapus', 'error');
+            return redirect('/kelolaPengguna');
+        }
     }
 }
